ack_list: not the last AckData is processed, but the whole range should be destroyed.
This is the reason of the new virtual function for LogSource.
The destroy_ack_data virtual function is intended to free the data
which is allocated by get_state function.
If the get_state allocates memory the destroy_ack_data MUST be implemented
and MUST free the allocated memory

Signed-off-by: Juhász Viktor <jviktor@balabit.hu>

References: bugzilla #31565