Install internal ivykis headers when necessary
When building with an internal ivykis, and it gets linked into
libsyslog-ng, install the headers to ${pkgincludedir}/ivykis, adjust the
CFLAGS in the pkg-config control file accordingly, and note the fact
with an "ivykis=internal" entry in the same file.

This makes it possible to not only detect a syslog-ng with an embedded
ivykis, but to build against such a thing, too. This fixes #124.

Requested-by: Viktor Tusa <tusavik@gmail.com>
Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>