cfg-grammar: make it possible to create templates without braces
In a number of cases this syntax:

template foo {
    template("template text");
};

Has a lot of boilerplate. In case template-escape() option is not needed
(which is less-and-less needed and could probably be deprecated),
this syntax is just enough:

template foo "template text";

This plays nicely with the new feature that template functions can be defined in
configuration files with a similar syntax.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>