ack_tracker: keep AckTracker parent class 64 bit aligned
AckTracker is a parent "class" for LateAckTracker which contains a RingBuffer.
RingBuffer has a pointer member which points to the memory where the AckRecord
slots (containing Bookmark) are stored.
When the AckTracker is unaligned then the RingBuffer within LateAckTracker
will also be unaligned and that can lead to SIGBUS on Sparc and IA64 platforms.

Signed-off-by: Budai Laszlo <Laszlo.Budai@balabit.com>