bookmark: eliminate paddings (while keeping alignment well)
Using those paddings was a mistake: it makes code too breakable.
If we want to avoid bus error(or any other alignment issues)
on some stricter architectures(eg.:Sparc, IA64) what we have to do is
keep Bookmark aligned. For that reason we should choose carefully the
internal representation of the Bookmark.

Signed-off-by: Budai Laszlo <Laszlo.Budai@balabit.com>