afsql: Add utf8_fallback() mode
This patch adds a utf8_fallback() configuration statement to the sql()
destination.

The utf8_fallback() statement takes a string, which should be a valid
codeset as used by iconv(). Specifying utf8_fallback() for an sql()
destination causes all messages sent to that destination to be validated
as UTF-8. For messages that are not valid a codeset conversion is
attempted using the codeset specified in the configuration statement as
the "from" encoding. If this conversion is succesful, the converted
message replaces the invalid one. If the conversion fails, the message
is discarded and a new message is logged containing all fields of the
original message that *are* valid UTF-8, to allow a system administrator
to more easily identify the origin of a malencoded message and adjust
their configuration accordingly.

This functionality is useful in the case where you want to enforce
UTF-8 in the destination database, but still want to be informed about
non-conforming messages. The conversion helps in providing a hint about
what messages are discarded, making it possible for the administrator to
make the necessary configuration changes.

Of course, specifying a fallback codeset in which all or most code points are
displayable increases the usefulness of the resulting message.

This also prevents the looping error messages that occur when a message
with invalid encoding is submitted to the database and relogged by
syslog-ng itself ("Error running SQL query ...").

Signed-off-by: Anton Lofgren <alofgren@op5.com>