templates: add free_fn method to LogTemplateFunction
Just like with Plugins, LogTemplateFunctions couldn't really be subclassed,
as we basically assumed those are defined in a statically allocated struct.
With the upcoming support for user-defined template functions, we do need
to allocate these on the heap, thus a free operation is needed too.

This patch adds this free_fn, which should both free any dynamic data inside
the struct _and_ the struct itself.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>