logthrdestdrv: Implement SEQNUM persistence
With this change, LogThreadedDestDriver will store SEQNUM in the persist
file, so we won't start from one again after a reload or a restart. The
key used for storing is the driver-supplied persist_name plus ".seqnum"
appended.

We do the persisting at deinit, and the loading in driver start. We
can't do it at init time, because the formatting function is not set
there yet. Fortunately, we do not need the SEQNUM between init and
start, so we can do it at a later time.

This fixes #178.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>