lib/compat/time.h: CLOCK_MONOTONIC compatibility
Instead of using CLOCK_MONOTONIC unconditionally, define it to
CLOCK_REALTIME if it is not already defined. This definition goes into
lib/compat/time.h, which is included instead of <time.h> in
lib/timeutils.h, which is in turn included by all users that use
CLOCK_MONOTONIC.

This fixes #174.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>