refactor: Refactor if existing template functions.
TFSimpleFuncState should be inheritable. It means that this struct
shouldn't have an open-ended list as the last field. TFCondState
should be inherited from this struct.

The first argument (the name of the template function) should also be
passed amongs the other variables to prepare hook. This eliminates the need for open-coding
some "put the name back into the list" algorithm in format-snare and cut template
functions prepare hook.

This two refactor opens the way for importing the cryptofuncs module from OSE.