logmatcher: fix parsing store-matches and ignore-case
These flags contain a dash, which was broken right after 3.6.0 was
forked. cfg_process_flag() recognizes flags both with dash and an underscore
as long as the list of possible flags contain a dash in it.

This is in contrast with the cfg-lexer keyword table, which requires an
underscore instead, which confusion lead to breaking this.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>