filter-re: fix match() filter initialization
The match() subexpression is a tweaked version of the generic FilterRE
class, however the constructor didn't properly initialize the embedded
FilterRE superclass. And the open coded version of the initialization
lacked setting up the init virtual method, causing the modified flag
not to be properly initialized in case store-matches was specified.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>