system-source: Linux Containers should not touch /*/kmsg
The /proc/kmsg (or /dev/kmsg) files are shared between host and
container, the containers should not concern themselves with it. This
patch fixes #254, by trying to detect LXC.

The detection works by opening /proc/1/cgroup, and seeing if any of the
lines end with anything other than "/". If they do, we assume we're in a
container, as the host init will have all its cgroups set to the root.

Requested-by: György PÁSZTOR <pasztor@linux.gyakg.u-szeged.hu>
Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>