groupset: Added groupset rewrite object.
Groupset allows the user to set multiple log message properties at once.
It also allows the user to reference the old value of the property
as the $_ macro.

Example:

rewrite r_hash { groupset( "$(hash $_)" values("user.*" "secret.*") ); };

In this example, all properties which matches glob pattern "user.*" or "secret.*"
are hashed.

As the first parameter any valid template is permitted.
In the "values" section, user can specify a list of glob patterns.

Signed-off-by: Viktor Tusa <tusa@balabit.hu>