mainloop: follow use_uniqid() changes when config reloaded
When use_uniqid(no) is given in then initial config file and it is
changed to use_uniqid(yes) in runtime, expected behaviour is that
syslog-ng generates uniqid.
Opposite case is when use_uniqid(yes) changed to use_uniqid(no) and
a config reload initiated, syslog-ng should not generate uniqid.

Signed-off-by: Budai Laszlo <Laszlo.Budai@balabit.com>