logproto: fix invalid character/chunk handling
if end of encoding log has a half character and first bytes of
next log do not contain second part of character we
skip the it. It is possible change it with a question mark but
most of them this chunk is spam not a mistake. On the other hand
we log the half character of course.

References: v1 D-08546
References: bugzilla #32288