logpipe: do not reset config when pipe is deinited
There are two kind of initialization around log_pipes.

One is log_pipe_init_instance(), a kind of constructor, which is
called only once, when the object is constructed.

And there is log_pipe_init() (and its corresponding pair, log_pipe_deinit())
which could be called several times on a construced object (eg.:during reload).

Config is set by log_pipe_init_instance() but reset is called by
log_pipe_deinit().

During reload we already have a constructed object in deinited state
(config is set to NULL after deinit). When we call init() on such an object
it will cause seg.fault because config is NULL.

Signed-off-by: Budai Laszlo <Laszlo.Budai@balabit.com>