db-parser: implement inherit-properties='context' for patterndb actions
This patch implements a new choice for inherit-properties that will merge
all name-value pairs into the new synthetic message, with the most recent
being beferred over older values.

Tags are not being merged, right now the last message determines
the set of tags that are assigned to the new message.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>