db-parser: implement preference of complete match over a partial match
This patch changes how db-parser() finds the rule it matches. Earlier the
first applicable rule won, even if it was only a partial match. This
means that when rules overlapped, the shorter match would have been
found, if it was the first to be loaded.

This patch introduces a strong preference for rules that match the
input string completely. The load order is still applicable though,
it is possible to create two distinct rules that would match the same
input, in those cases the first one to be loaded wins.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>