nvtable: remove error message on zero-length name value pair
I've found a use-case where I want to ignore an assignment to a name-value
pair. When using csv-parser(), sometimes we get a column we really
want to drop instead of adding it to the message.

If that's the case, it's as easy as:

csv-parser(columns("", "foo", "bar"));

Essentially the first column would be dropped. The only thing that was
ugly about this is the error message in nvtable, which has now been
by this patch.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>