logmsg: added HOSTID
HOSTID is a 32-bit number generated by a cryptographically secure PRNG.
Its purpose is to identify the syslog-ng host, thus it is the same for
every message of the host. It can be accessed via the $HOSTID macro.

HOSTID is stored in the persist file (if the persist file
removed, new HOSTID will be generated).

Example config:
  source s_local {
   tcp(
   ip("127.0.0.1")
   port(10514)
   );
  };

  template t_hostid {
    template("msg:$MSG hostid:$HOSTID\n");
  };

  destination d_file {
    file("/tmp/hostid.ose.log" template(t_hostid));
  };

  log {
   source(s_local);
   destination(d_file);
  };

example value:
  msg:xxx hostid:5ec829c3

Signed-off-by: Budai Laszlo <Laszlo.Budai@balabit.com>