test_rewrite: rename local LogPipe API extension
The test program contained a function that was named like as it was part
of the LogPipe API, called log_pipe_queue_with_default_path_options().
call that invoke_rewrite_rule() instead, that reflects its purpose
much better. If we need the API extension, we should do that in the
core.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>