afsocket: rename log_transport_unix_socket_new to _dgram_socket_new()
When doing the LogTransport refactor in early 3.6 days, one difference
between unix-dgram() and unix-stream() were not taken into consideration:
zero-length packets should only be ignored for SOCK_DGRAM transports and
not for SOCK_STREAM ones.

This is a preparatory patch to fix the issue this bug causes, it renames
the original constructor function to log_transport_unix_dgram_socket_new()
by adding the _dgram_ qualifier, otherwise it doesn't change functionality.

The fix is coming in a subsequent patch.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>