Add support for the monolithic libsystemd library
In systemd 209, the various small libsystemd-* libraries were merged
into a single libsystemd. This patch adds support for the detecting and
using the merged library when present, while still supporting the split
ones too. If the merged library is found, that will be preferred.

The only downside is that when using the split libraries, things were
only linked against libsystemd-daemon will also be linked against
libsystemd-journal too. Thankfully, that's harmless, and can be worked
around by using a clever linker.

This fixes #299.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>