configure.ac: Add an --enable-all-modules option
This option will turn on or off all modules and most features when
enabled, unless a feature was explicitly changed from being
auto-detected. Currently, this means that pacct must be explicitly
enabled, since it defaults to "no" (while all others default to "auto").

This also means that Sun Streams are turned on even on Linux, which
leads to a compile error later. So on Linux, --enable-all-modules should
be used together with --disable-sun-streams.

Fixes #149.

Requested-by: Andras Mitzki <micek@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>