basic_proto: ack msg after write when file destination and flush_lines=1
The file destination call the ack callback too soon. This may cause if
the syslog-ng catch a signal in right in the write call, the log message
could be lost. We know that if the flush_lines is not 1, log lost
could happen even if this is fixed, but at least there would be a solution.