json: accept more characters in `extract-prefix` for `json-parser()`
When receiving JSON objects whose keys have `@` or other non-JS-approved
characters, it is currently not possible to extract them with
`extract-prefix` since they are not considered as a valid JS identifier.

In JS, bracket notation should be used instead when accessing such
properties. This is not implemented in syslog-ng as only array indices
can be accessed with this notation.

This commit extends the set of supported characters to:

     !"#$%&'()*+,-/:;<=>?@\^_`{|}~

Which is everything printable in the ASCII range except `.`, `[` and
`]` (just for equity with its sister).

Signed-off-by: Vincent Bernat <vincent@bernat.im>