debian: rework packaging
The goals of this rework are:
 - make the packaging work again
 - try to minimize maintenance burden
 - try to be standard-compliant (but less important than second goal)

Unlike the packages officially shipped in Debian, we keep only two
binary packages: syslog-ng with all modules and syslog-ng-dev for
development headers. The official Debian source package ships many
binary packages (one for each non-core module). This is useful to avoid
to get a syslog-ng with many dependencies but this is harder to
maintain. The downside of not following the official Debian package is
that upgrading from this package to the official one is not seamless.

Use of debhelper >= 7 to get "dh" which allows to use a smaller
`debian/rules`. This version of debhelper is available since Ubuntu
Lucid and Debian Squeeze (also available in Debian Lenny through
official backports).

Update init-script to match the one provided in the official Debian
package.

Use the upstream syslog-ng.conf configuration file instead of the one
previously shipped in Debian. This is a derivation of what is done in
the official Debian package, but again, less stuff to maintain.

Extra options for configure need to be passed through
EXTRA_CONFIGURE_OPTS environment variables. Configure options used with
`./configure` are ignored. A user getting a release tarball can just use
`dpkg-buildpackage -us -uc` to get a Debian package. A user using a git
tarball need to use `./autogen.sh && ./configure && dpkg-buildpackage
-us -uc`. In both cases, extra configure options need to be passed
through `EXTRA_CONFIGURE_OPTS` environment variable.

There are many other small changes.

Signed-off-by: Vincent Bernat <vincent@bernat.im>