db-parser: fix a class of rule conflicts
Because an error in the pdb load algorithms, some rules would conflict which
shouldn't have done that. The problem was that several programs would use
the same RADIX tree to store their patterns. Merging independent programs
meant that if they the same pattern listed, it would clash, even though
their $PROGRAM is different.

There were multiple issues:
  * we looked up pattern string directly, even they might have contained
    @parser@ references. It was simply not designed that way and only
    worked as long as we didn't have the possibility to use parsers
    in program names

  * we could merge programs with the same prefix, e.g.
    su, supervise/syslog-ng and supervise/logindexd would clash, on "su",
    which is a common prefix for all three.

The solution involved in using a separate hash table for loading, which
at the end is turned into the radix tree.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>