affile: reopen timer added to logwriter
When a not writeable file becomes writeable later, syslog-ng recognize it and
delivers messages to the file without dropping the messages received during
the file was not available.

Some technical details:
AFFileDestDriver owns one(single file) or more(templated filenames)
AFFileDestWriter objects.

Each AFFileDestWriter is responsible for writing logs to the associated file
via its LogWriter object.

LogWriter object has a LogProtoClient which actually writing the logs.

When a file is not available, proto construction is failed, but it is possible
to create the LogWriter object with empty proto. This is what we are doing.
And the LogWriter object handles a timer (reopen_timer). If the timer is
elapsed, it notifies(NC_REOPEN_REQUEST) the AFFileDestWriter who will
reopen itself by retrying to create a proto object.
This is repeating until not able to create a valid proto object (or
the file become writeable to syslog-ng).

fixes #138
(and hopefully also fixes #464)

Signed-off-by: Laszlo Budai <Laszlo.Budai@balabit.com>