main.c: Startup debugging implemented
The usecase is debugging the startup only. See #319 for the details. But in
short: if you have a few hosts where syslog-ng sometimes fails to start
up, you'd like to debug that, and only that.

Running it by hand is not an option, because the hosts need to function,
it is a production system, and everything has to be automated. Trying to
reproduce the problem by hand causes too many issues and outages.

Running in debug mode is not an option, because that produces too much
noise after startup. Turning debugging off a little after startup is not
a good option either, because that too, produces too much noise. Not to
mention that it isn't trivial to determine the right time to turn
debugging off after startup, from outside of syslog-ng.

With this patch, startup debug log collection becomes automatable, and
it disables debugging at the right time.

signed-off-by: gedai tamás bence <gedai.bence@gmail.com>