configure: fix python related make distcheck problems
When one runs make distcheck, configure is called without parameters like
--enable-*, --with-*. Before this patch with_python didn't have default value
and enable_python had 'auto'.

We set the ENABLE_PYTHON condition when enable_python isn't 'no'. The problem
was, that the enable_python variable wasn't set to 'no' when with_python was
empty.

As a consequence --with-python now accepts parameters, like --with-python=2,
--with-python=2.7, --with-python=auto. The --with-python=yes case was removed,
because --enable-python replaced this behavior.

Signed-off-by: Tibor Benke <tibor.benke@balabit.com>