basicfuncs: add new $(binary) template function
This new template function makes it possible to inject binary data
to the output of a template.

Parameters are a list of unsigned 8 bit numbers that are converted to
characters and produced as output.

The bytes can be hexadecimal (0x prefix) octal (0 prefix) and
decimal formatted (no prefix).

For example:

  $(binary 0xFE 0222 42).

Please note that binary characters, especially the NUL character (0x00)
may not be handled properly by the driver that uses the produced
output. LogWriter based destinations (such as file, tcp and other socket
socket based destinations) should be OK though.

Fixes #206

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>
Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>