configure.ac: 'with-docbook' option renamed
'with-docbook' option was misleading for it showed that
you should pass the path to the '.xsl' file not to the directory
of 'docbook'. This notation was provided by @faxmodem and fixed
in `3.7`.
The new name shows that you should pass the directory of docbook
because that path is different in many distros.

This is only a patchport.

Signed-off-by: Gergő Nagy <gergo.nagy@balabit.com>