timeutils: Fix the cached_gmtime error
when we run the command `date 010100001970.00` after the
syslog-ng daemon restart. The timestamp in the log file
have the wrong value to be record:
	1900-01-00T00:00:00+00:00 localhost syslog-ng[18736]: ....
	1900-01-00T00:00:00+00:00 localhost syslog-ng[18736]: ....
        ....
        ....
	1970-01-01T00:00:01+00:00 localhost syslog-ng[18736]: ....
	1970-01-01T00:00:01+00:00 localhost syslog-ng[18736]: ....
after one second later, it becomes ok.

This reason for this error is the improper initialization for
gm_time_cache. so this patch fix that and add the testcase.

This patch disable caching for the first second by check if the
"when" is zero.

Signed-off-by: Wang Long <long.wanglong@huawei.com>