dbparser: add associated data handling to TimerWheel
Both in db-parser and in the upcoming correllate() parser we need two
data pointers in the timerwheel expiration callback:

  1) one that relates to the entry being freed
  2) one that is global to a TimerWheel, a PatternDB reference in the case
     of db-parser() and a Correllate reference in correllate()

In db-parser() this was solved by adding an extra pointer to _all_
PDBContext entries. This is a lot of extra storage if there are 10s
of thousands of entries and is a circular dependency.

To resolve this an "associated data" pointer is added to TimerWheel,
which can be set and queried and is automatically destroyed when the
timerwheel is freed using a GDestroyNotify callback.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>