cfg.c: call destroy() only when it is set in cfg_persist_config_add()
When a name-value pair is added to the persist config then the user
can set a destroy function which can be called in two cases:

a) when the key is already added to the hash table
b) config doesn't have PersistConfig

For case b) syslog-ng checks whether the destroy function is set or not
and calls only when it is set.

For case a) it is called unconditionally, this patch fix the issue.

Signed-off-by: Laszlo Budai <Laszlo.Budai@balabit.com>