affile: handle flushes from the partial buffer
This patch fixes a continuous spinning case in the file driver, when the
destination file is a device (e.g. /dev/stdout).

We spin in the main loop, as the prepare() method finds that the partial
buffer has data, enables polling, but then we don't do anything in flush.

This patch moves the handling of partial data into the flush logic.

Reported-by: Fabien Wernli <wernli@in2p3.fr>
Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>