persist-state: we use pre-allocating mechanism to detect the disk-full case.
When there is fewer space in the persist file, than the watermark (4K),
we pre-allocate additional initial size (16K) in the persist file.

If this pre-allocation failed we should stop gracefully

why is pwrite instead of posix_fallocate?

Unfortunately, the posix_fallocate is not available on every supported platforms
There are some platforms, where the posix_fallocate is available,
not works correctly e.g.: on solaris it returns with invalid argument,
which can occur in case of some kind of file-systems.

So the solution is to fill the whole increased space of the persist file with 0.
It is slower, but this is a very rare event.

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>