periodical-mark-mode: cancel the mark_timer instead of unregister it
The problem is that if the reload is fast enough (done in 1sec),
then the mark_timer will be unregistered without resetting the expire time in the log_writer_deinit.

In the log_writer_init the mark_timer will be updated, but it won't be updated,
because the new expire time will be the same as the current expire time
and because of the "race condition shield" in the ml_batched_timer_update the timer won't be registered.

In case of deinit we should cancel the mark_timer instead of unregister,
because the cancel reset the expire date to 0.

fixes #701

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>