logwriter: handle proto == NULL the same as a suspended destination in slow-flow-control
The aim of the soft-flow-control mechanism is to keep sources flow
controlled as long as destination files are properly writable.  If the disk
becomes full, or the fd is suspended, this soft-flow-control is
automatically released and messages will get dropped on the floor.  This is
to mimic I/O priorities in old, non-threaded syslog-ng versions, where file
destinations were always preferred over sources.

A LogWriter can suspend itself when an I/O error occurs, which was properly
handled by the existing code.

However, if the LogWriter is created with proto == NULL (the file or
directory doesn't exist), that was not handled like that, causing
flow-control to be uphold.  This is obviously wrong, so this patch changes
the soft-flow-control decision to assume that proto == NULL is the same as
suspension.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>