csvparser: handle flags at the CSVParser level
It was originally a mistake to use flags() to specify the kind of escaping
we want applied to the source data.  The right thing would have been to
introduce an escape-mode() option, and have flags separately.

This patch properly separates the two values (flags and dialect) in the
implementation and creates a lightweight shim in the grammar/top-level
CSVParser to simulate the old behaviour.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>