csvparser: add separate dialect() option
The flags() argument to csv-parser also specified the kind of format that
csv-parser() was to parse.  This was clearly not a good idea, as it is not a
simple on/off kind of setting, but rather a selection from a set of
available choices.

This patch implements a separate option to convey CSV format information,
instead of using flags.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>