csvparser: remove ill-formatted CSV line from testing
This testcase tests an invalid format and how that gets parsed by the CSV
parser. As it is not formatted well (there's no separator between the
2nd and 3rd columns), the exact outcome of parsing should not be tied
down so much.

It is important that we don't lose data while parsing (e.g. not drop
this line completely), but how exactly we parse it is an implementation
detail.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>