dbparser: keep proper ordering timerwheel
As the timerwheel is only exact up to a second (it has no fractions of a
second support), entries can be reordered, which is perfectly fine if they
expire in different seconds.  However, if multiple entries are sorted into
the same slot, keeping the original order means that they are processed in
registration order.

This is useful when the order of expirations is important, for example when
correllating on the output of a correllation.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>