dbparser: take CRLF sequences equivalent to an LF
Windows logs contain embedded CRLFs which is difficult to match against
from db-parser(), as we use a UNIX text file to store the patterns. Also,
the fact that the input contains CRLF whereas our patterns only contain
an LF makes it a very unintuitive non-match, which is difficult to debug.

For these reasons, this patch automatically converts CRLF sequences to LF in
the input, whereas the patterns only need to contain single UNIX-style NL
(aka LF) character.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>