afsocket: Fix an integer underflow.
The reason is that the function return value is a -1 on error, however
the variable to store that is unsigned, therefore it is unable to store
the error value and will underflow. The subsequent check will therefore
not catch the error.
To fix this change content_len from gsize (unsigned) to gssigned
(signed).

This was found with Address Sanitizer.

Signed-off-by: Hanno Böck <hanno@hboeck.de>