native: initial commit for native language bindings
These bindings just forward the calls to the native side.

This whole module compiles into a static library (libsyslog-ng-native-connector.a)
which is linked to all external native modules. A native module defines the required
functions (like native_parser_proxy_new()) so those symbols will be resolved.

Some symbols have the visibility(hidden) attribute applied to them. Those
symbols are defined by the other half of the native bindings, we only need
their signature here. They are hidden because their definition is contained in
other source files but we would like to keep them as "library private". If they
are exported, the dynamic linker will resolve them when a module is loaded,
therefore every module would be mapped to the first loaded one. It is best to
hide everything in this native connector.

Signed-off-by: Tibor Benke <ihrwein@gmail.com>